requirements.txt
# Core dependencies
praw>=7.8.1
prawcore>=2.4.0,<3.0.0
update-checker>=0.18.0

# Web framework
streamlit>=1.31.1

# AI integration
google-generativeai>=0.3.2

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
tqdm>=4.66.1
requests>=2.31.0
urllib3>=2.0.0

# For Streamlit deployment
pandas>=2.0.0
numpy>=1.24.0
